Set initial deadline to 15s
Currently, a client could open a connection and then never send any
requests or complete authentication for an indefinite amount of time.

This patch sets a deadline for the initial handshake and command of 15s,
so that if the client did not request a dial before then, the connection
is closed.